Added wrapper around geteuid() to check if process needs elevation
